# apps/client/latest_openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
...
    "/daily-snippets/{snippet_id}": {
      "get": {
        "tags": ["daily-snippets"],
        "summary": "Get Daily Snippet",
        "operationId": "get_daily_snippet_daily_snippets__snippet_id__get",
        "parameters": [
          {
            "name": "snippet_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Snippet Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailySnippetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["daily-snippets"],
        "summary": "Delete Daily Snippet",
        "operationId": "delete_daily_snippet_daily_snippets__snippet_id__delete",
        "parameters": [
          {
            "name": "snippet_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Snippet Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/daily-snippets": {
      "get": {
        "tags": ["daily-snippets"],
        "summary": "List Daily Snippets",
        "operationId": "list_daily_snippets_daily_snippets_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 50, "title": "Limit" }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Offset" }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "desc", "title": "Order" }
          },
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "From Date"
            }
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "To Date"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailySnippetListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["daily-snippets"],
        "summary": "Create Daily Snippet",
        "operationId": "create_daily_snippet_daily_snippets_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DailySnippetCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailySnippetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/weekly-snippets/{snippet_id}": {
      "get": {
        "tags": ["weekly-snippets"],
        "summary": "Get Weekly Snippet",
        "operationId": "get_weekly_snippet_weekly_snippets__snippet_id__get",
        "parameters": [
          {
            "name": "snippet_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Snippet Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklySnippetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["weekly-snippets"],
        "summary": "Delete Weekly Snippet",
        "operationId": "delete_weekly_snippet_weekly_snippets__snippet_id__delete",
        "parameters": [
          {
            "name": "snippet_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Snippet Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/weekly-snippets": {
      "get": {
        "tags": ["weekly-snippets"],
        "summary": "List Weekly Snippets",
        "operationId": "list_weekly_snippets_weekly_snippets_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 50, "title": "Limit" }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Offset" }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "desc", "title": "Order" }
          },
          {
            "name": "from_week",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "From Week"
            }
          },
          {
            "name": "to_week",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "To Week"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklySnippetListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["weekly-snippets"],
        "summary": "Create Weekly Snippet",
        "operationId": "create_weekly_snippet_weekly_snippets_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WeeklySnippetCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklySnippetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DailySnippetCreate": {
        "properties": { "content": { "type": "string", "title": "Content" } },
        "type": "object",
        "required": ["content"],
        "title": "DailySnippetCreate"
      },
      "DailySnippetListResponse": {
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/DailySnippetResponse" },
            "type": "array",
            "title": "Items"
          },
          "total": { "type": "integer", "title": "Total" },
          "limit": { "type": "integer", "title": "Limit" },
          "offset": { "type": "integer", "title": "Offset" }
        },
        "type": "object",
        "required": ["items", "total", "limit", "offset"],
        "title": "DailySnippetListResponse"
      },
      "DailySnippetResponse": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "user_id": { "type": "integer", "title": "User Id" },
          "date": { "type": "string", "format": "date", "title": "Date" },
          "content": { "type": "string", "title": "Content" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "date",
          "content",
          "created_at",
          "updated_at"
        ],
        "title": "DailySnippetResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "WeeklySnippetCreate": {
        "properties": { "content": { "type": "string", "title": "Content" } },
        "type": "object",
        "required": ["content"],
        "title": "WeeklySnippetCreate"
      },
      "WeeklySnippetListResponse": {
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/WeeklySnippetResponse" },
            "type": "array",
            "title": "Items"
          },
          "total": { "type": "integer", "title": "Total" },
          "limit": { "type": "integer", "title": "Limit" },
          "offset": { "type": "integer", "title": "Offset" }
        },
        "type": "object",
        "required": ["items", "total", "limit", "offset"],
        "title": "WeeklySnippetListResponse"
      },
      "WeeklySnippetResponse": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "user_id": { "type": "integer", "title": "User Id" },
          "week": { "type": "string", "format": "date", "title": "Week" },
          "content": { "type": "string", "title": "Content" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "week",
          "content",
          "created_at",
          "updated_at"
        ],
        "title": "WeeklySnippetResponse"
      }
    }
  }
}